requirements.txt
tigeropen>=1.0.0
pandas>=1.5.0
python-dotenv>=0.19.0
tenacity>=8.0.0

# Observability & Polish (Phase 8)
prometheus-client>=0.18.0
python-json-logger>=2.0.0
fastapi>=0.95.0
uvicorn[standard]>=0.21.0

# Database & Migrations
sqlalchemy>=2.0.0
alembic>=1.10.0

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
